Move route handlers to "routes" folder
